split string stuff and assert
