+++ vs/mob.vcxproj
@@ -106,6 +106,7 @@
     <ClCompile Include="..\src\tools\process_runner.cpp" />
     <ClCompile Include="..\src\tools\tools.cpp" />
     <ClCompile Include="..\src\utility.cpp" />
+    <ClCompile Include="..\src\utility\string.cpp" />
     <ClCompile Include="..\src\utility\threading.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -121,6 +122,8 @@
     <ClInclude Include="..\src\tasks\tasks.h" />
     <ClInclude Include="..\src\tools\tools.h" />
     <ClInclude Include="..\src\utility.h" />
+    <ClInclude Include="..\src\utility\assert.h" />
+    <ClInclude Include="..\src\utility\string.h" />
     <ClInclude Include="..\src\utility\threading.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
