+++ vs/mob.vcxproj.filters
@@ -156,6 +156,9 @@
     <ClCompile Include="..\src\utility\threading.cpp">
       <Filter>src\utility</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\utility\string.cpp">
+      <Filter>src\utility</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\pch.h">
@@ -195,6 +198,12 @@
       <Filter>src\core</Filter>
     </ClInclude>
     <ClInclude Include="..\src\utility\threading.h">
+      <Filter>src\utility</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\utility\string.h">
+      <Filter>src\utility</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\utility\assert.h">
       <Filter>src\utility</Filter>
     </ClInclude>
   </ItemGroup>
